+++ hardhat-config/networks.config.js
@@ -47,20 +47,6 @@ module.exports = {
     },
 
     // Staging
-    goerli: {
-      // chainId 5
-      url: 'http://goerli',
-      live: true,
-      tags: ['staging', 'ethereum'],
-      companionNetworks: {
-        polygon: 'mumbai',
-        bsc: 'bsctest',
-        okex: 'okextest',
-        arbitrum: 'arbgoerli',
-      },
-    },
-
-    // Staging
     sepolia: {
       // chainId 11155111
       url: 'http://sepolia',
@@ -87,18 +73,6 @@ module.exports = {
       tags: ['production', 'polygon'],
       companionNetworks: {
         ethereum: 'mainnet',
-      },
-    },
-
-    // Staging
-    mumbai: {
-      // chainId 80001
-      // https://static.matic.network/network/testnet/mumbai/index.json
-      url: 'http://mumbai',
-      live: true,
-      tags: ['staging', 'polygon'],
-      companionNetworks: {
-        ethereum: 'goerli',
       },
     },
 
@@ -164,7 +138,7 @@ module.exports = {
       live: true,
       tags: ['staging', 'okex'],
       companionNetworks: {
-        ethereum: 'goerli',
+        ethereum: 'sepolia',
       },
     },
 
@@ -181,17 +155,6 @@ module.exports = {
       companionNetworks: {
         ethereum: 'mainnet',
         bsc: 'bsc',
-      },
-    },
-
-    arbgoerli: {
-      // chainId 421613
-      // https://developer.arbitrum.io/public-chains
-      url: 'http://arbgoerli',
-      live: true,
-      tags: ['staging', 'arbitrum'],
-      companionNetworks: {
-        ethereum: 'goerli',
       },
     },