+++ hardhat-config/providers.config.js
@@ -2,29 +2,22 @@ module.exports = {
   providers: {
     infura: {
       mainnet: '{{PROVIDER_URL_INFURA_MAINNET}}',
-      goerli: '{{PROVIDER_URL_INFURA_GOERLI}}',
       sepolia: '{{PROVIDER_URL_INFURA_SEPOLIA}}',
       matic: '{{PROVIDER_URL_INFURA_MATIC}}',
-      mumbai: '{{PROVIDER_URL_INFURA_MUMBAI}}',
       amoy: '{{PROVIDER_URL_INFURA_AMOY}}',
       arb1: '{{PROVIDER_URL_INFURA_ARB1}}',
-      arbgoerli: '{{PROVIDER_URL_INFURA_ARBGOERLI}}',
       arbsepolia: '{{PROVIDER_URL_INFURA_ARBSEPOLIA}}',
     },
     alchemy: {
       mainnet: '{{PROVIDER_URL_ALCHEMY_MAINNET}}',
-      goerli: '{{PROVIDER_URL_ALCHEMY_GOERLI}}',
       sepolia: '{{PROVIDER_URL_ALCHEMY_SEPOLIA}}',
       matic: '{{PROVIDER_URL_ALCHEMY_MATIC}}',
-      mumbai: '{{PROVIDER_URL_ALCHEMY_MUMBAI}}',
       amoy: '{{PROVIDER_URL_ALCHEMY_AMOY}}',
       arb1: '{{PROVIDER_URL_ALCHEMY_ARB1}}',
-      arbgoerli: '{{PROVIDER_URL_ALCHEMY_ARBGOERLI}}',
       arbsepolia: '{{PROVIDER_URL_ALCHEMY_ARBSEPOLIA}}',
     },
     maticvigil: {
       matic: '{{PROVIDER_URL_MATICVIGIL_MATIC}}',
-      mumbai: '{{PROVIDER_URL_MATICVIGIL_MUMBAI}}',
     },
     binance: {
       bsc: 'https://bsc-dataseed.binance.org/',
@@ -36,7 +29,6 @@ module.exports = {
     },
     arbitrum: {
       arb1: 'https://arb1.arbitrum.io/rpc',
-      arbgoerli: 'https://goerli-rollup.arbitrum.io/rpc',
       arbsepolia: 'https://sepolia-rollup.arbitrum.io/rpc',
     },
     sepolia: {