+++ CHANGELOG.md
@@ -7,9 +7,12 @@
 - Added networks and providers configs for Arbitrum Sepolia network.
 - Added arbitrum networks as companion networks of bsc networks.
 
-### Breaking changes
+### Improvements
 - Updated to dependency version `@animoca/ethereum-contract-helpers@1.0.3`.
+
+### Breaking changes
 - Updated to use latest testnets(amoy, arbsepolia) instead of deprecated testnets(mumbai, arbgoerli) as companion networks.
+- Removed deprecated network configs including goerli, mumbai, arbgoerli.
 
 ## 3.0.3